React app: add react router
