+++ .tps/react-app/default/package.json.tps
@@ -5,6 +5,7 @@
 	"dependencies": {
 		"react": "^18.3.1",
 		"react-dom": "^18.3.1",
+		"react-router-dom": "^6.28.0",
 		"react-scripts": "5.0.1",
 		"web-vitals": "^2.1.4"
 	},
