+++ .tps/react-app/default/src/routes/Home/Home.css
@@ -1,19 +1,19 @@
-.App {
+.Home {
 	text-align: center;
 }
 
-.App-logo {
+.Home-logo {
 	height: 40vmin;
 	pointer-events: none;
 }
 
 @media (prefers-reduced-motion: no-preference) {
-	.App-logo {
-		animation: App-logo-spin infinite 20s linear;
+	.Home-logo {
+		animation: Home-logo-spin infinite 20s linear;
 	}
 }
 
-.App-header {
+.Home-header {
 	background-color: #282c34;
 	min-height: 100vh;
 	display: flex;
@@ -24,11 +24,11 @@
 	color: white;
 }
 
-.App-link {
+.Home-link {
 	color: #61dafb;
 }
 
-@keyframes App-logo-spin {
+@keyframes Home-logo-spin {
 	from {
 		transform: rotate(0deg);
 	}
