+++ .tps/react-app/default/src/routes/Home/Home.js
@@ -1,0 +1,25 @@
+import logo from '../../logo.svg';
+import './Home.css';
+
+function Home() {
+	return (
+		<div className="Home">
+			<header className="Home-header">
+				<img src={logo} className="Home-logo" alt="logo" />
+				<p>
+					Edit <code>src/routers/Home.js</code> and save to reload.
+				</p>
+				<a
+					className="Home-link"
+					href="https://reactjs.org"
+					target="_blank"
+					rel="noopener noreferrer"
+				>
+					Learn React
+				</a>
+			</header>
+		</div>
+	);
+}
+
+export default Home;
